backend/requirements.txt
SQLAlchemy~=2.0.43
pydantic~=2.11.7
fastapi~=0.116.1
python-dotenv~=1.1.1
python-jose~=3.5.0
passlib[argon2]~=1.7.4
